Update Calculator.java
everything is doubles now!

+++ Calculator.java
@@ -3,7 +3,7 @@ public class Calculator {
 		return a + b;
 	}
 
-	public static int subtract(int a, int b){
+	public static double subtract(double a, double b){
 		return a - b;
 	}
 }
